Change namespaces for moved files
