+++ Homework/TagsCloudContainer/CloudPainter.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Drawing;
 using System.Drawing.Imaging;
+using TagsCloudContainer.PaintConfigs;
+using TagsCloudContainer.TextParsers;
 using TagsCloudVisualization;
 
 namespace TagsCloudContainer
