+++ Homework/TagsCloudContainer/PaintConfigs/PaintCofig.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 
-namespace TagsCloudContainer
+namespace TagsCloudContainer.PaintConfigs
 {
     public class PaintCofig : IPaintConfig
     {
