+++ Homework/TagsCloudContainer/TextParsers/ITextParser.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace TagsCloudContainer
+namespace TagsCloudContainer.TextParsers
 {
     public interface ITextParser
     {
